Move the advanced APIs to Internal module for now
To keep the exported interface simpler.
